Refactor Bible core module structure
