+++ crates/presenter-core/src/bible/mod.rs
@@ -1,0 +1,16 @@
+mod canonical;
+mod reference;
+mod search;
+mod translation;
+
+pub use canonical::{
+    canonical_book_by_code, canonical_book_by_name, canonical_book_by_number, BibleBookCanonical,
+};
+pub use reference::{BibleReference, BibleReferenceError};
+pub use translation::{
+    BibleBookChapterSummary, BibleBroadcast, BibleIngestionBatch, BibleIngestionError,
+    BiblePassage, BiblePreferences, BiblePreferencesDraft, BibleTranslation,
+};
+
+#[cfg(test)]
+mod tests;
