+++ crates/presenter-core/src/bible/tests/mod.rs
@@ -1,0 +1,3 @@
+mod canonical;
+mod reference;
+mod translation;
